+++ premium-core-engine/src/main/java/adrianmikula/jakartamigration/advancedscanning/domain/SerializationCacheScanResult.java
@@ -1,7 +1,6 @@
 package adrianmikula.jakartamigration.advancedscanning.domain;
 
 import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -14,7 +13,6 @@ import java.util.List;
  */
 @Getter
 @NoArgsConstructor(force = true, access = AccessLevel.PRIVATE)
-@AllArgsConstructor
 public class SerializationCacheScanResult {
     private final String filePath;
     private final List<SerializationCacheUsage> usages;